Move RecoveryId to own module
