+++ src/lib.rs
@@ -24,41 +24,14 @@ mod scalar;
 pub use error::Error;
 
 mod keys;
+mod recovery_id;
 pub mod signature;
 pub mod util;
-pub mod message;
+mod message;
 
 pub use ecdh::SharedSecret;
 pub use keys::{PublicKey, SecretKey};
 pub use signature::{recover, verify, Signature};
 pub use message::Message;
+pub use recovery_id::RecoveryId;
 
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
-/// Tag used for public key recovery from signatures.
-pub struct RecoveryId(u8);
-
-impl RecoveryId {
-    pub fn parse(p: u8) -> Result<RecoveryId, Error> {
-        if p < 4 {
-            Ok(RecoveryId(p))
-        } else {
-            Err(Error::InvalidRecoveryId)
-        }
-    }
-
-    pub fn serialize(&self) -> u8 {
-        self.0
-    }
-}
-
-impl Into<u8> for RecoveryId {
-    fn into(self) -> u8 {
-        self.0
-    }
-}
-
-impl Into<i32> for RecoveryId {
-    fn into(self) -> i32 {
-        self.0 as i32
-    }
-}
