+++ src/recovery_id.rs
@@ -1,0 +1,31 @@
+use Error;
+
+#[derive(Debug, Clone, Copy, Eq, PartialEq)]
+/// Tag used for public key recovery from signatures.
+pub struct RecoveryId(pub u8);
+
+impl RecoveryId {
+    pub fn parse(p: u8) -> Result<RecoveryId, Error> {
+        if p < 4 {
+            Ok(RecoveryId(p))
+        } else {
+            Err(Error::InvalidRecoveryId)
+        }
+    }
+
+    pub fn serialize(&self) -> u8 {
+        self.0
+    }
+}
+
+impl Into<u8> for RecoveryId {
+    fn into(self) -> u8 {
+        self.0
+    }
+}
+
+impl Into<i32> for RecoveryId {
+    fn into(self) -> i32 {
+        self.0 as i32
+    }
+}
